Decrease Matplotlib figure DPI
Firefox has a bug where it does not show large PNGs that are
embedded in SVGs as data URLs on the first page load.

Decrease the figure DPI to work around this. The reduced resolution
should not be noticeable for most of the Matplotlib-generated images
in the User Guide because most embedded PNGs are gentle gradients.

Fixes #291.